+++ src/main/java/com/github/rjeschke/cetoneasm/actions/CounterDecrementAction.java
@@ -23,10 +23,10 @@ import com.github.rjeschke.cetoneasm.FileLocation;
 
 public class CounterDecrementAction extends Action
 {
-    private final long id;
-    private final long jumpId;
+    private final int id;
+    private final int jumpId;
 
-    public CounterDecrementAction(final FileLocation location, final long id, final long jumpId)
+    public CounterDecrementAction(final FileLocation location, final int id, final int jumpId)
     {
         super(location);
         this.id = id;
